+++ src/test/java/sehee/testutil/stub/NumberMakerStub.java
@@ -1,4 +1,4 @@
-package sehee.stub;
+package sehee.testutil.stub;
 
 import sehee.util.numbermaker.NumberMaker;
