+++ src/test/java/sehee/testutil/stub/ReaderStub.java
@@ -1,4 +1,4 @@
-package sehee.stub;
+package sehee.testutil.stub;
 
 import sehee.io.in.Reader;
